+++ fleche/stats.mli
@@ -8,5 +8,10 @@ end
 
 val get : kind:Kind.t -> float
 val record : kind:Kind.t -> f:('a -> 'b) -> 'a -> 'b * float
-val dump : unit -> string
+val to_string : unit -> string
 val reset : unit -> unit
+
+type t
+
+val dump : unit -> t
+val restore : t -> unit